backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
PyJWT>=2.8
bcrypt>=4.1
google-generativeai>=0.8
pypdf>=5.0
reportlab>=4.2
qrcode>=7.4
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=22.0
